creating flask blueprints so the main file is more manageable and fixing bad hotel link
